lab-8初始化

+++ include/fs/bitmap.h
@@ -4,7 +4,7 @@
 uint32 bitmap_alloc_block();
 uint32 bitmap_alloc_inode();
 void   bitmap_free_block(uint32 block_num);
-void   bitmap_free_inode(uint32 inode_num);
+void   bitmap_free_inode(uint16 inode_num);
 void   bitmap_print(uint32 bitmap_block_num);
 
 #endif
